### 2022-01-30 #4112 @whedon
[PRE REVIEW]: NGSTrefftz: Add-on to NGSolve for Trefftz methods
**Submitting author:** @PaulSt (<a href="http://orcid.org/0000-0001-5073-3366">Paul Stocker</a>)
**Repository:** <a href="https://github.com/PaulSt/NGSTrefftz" target ="_blank">https://github.com/PaulSt/NGSTrefftz</a>
**Version:** v0.0.0
