### 2022-02-01 #4112 @Nikoleta-v3
Hey 👋🏻 @PaulSt thank you again for your submission with JOSS, could you please:

- 1️⃣  take a look through https://bit.ly/joss-reviewers to suggest some potential reviewers
- 2️⃣  check and fix the missing DOIs: https://github.com/openjournals/joss-reviews/issues/4112#issuecomment-1025166198
