+++ doc/paper/paper.bib
@@ -84,11 +84,12 @@
   number={5},
   pages={316--337},
   year={2005},
-  publisher={American Society of Mechanical Engineers}
+  publisher={American Society of Mechanical Engineers},
+  doi={10.1115/1.1995716}
 }
 
 @Article{	  kk95,
-  title	        = "Trefftz method: an overview ",
+  title	        = "Trefftz method: an overview",
   sjournal       = "Adv. Eng. Softw.",
   journal      = "Advances in Engineering Software ",
   volume        = "24",
@@ -97,7 +98,8 @@
   year	        = "1995",
   note	        = "",
   issn	        = "0965-9978",
-  author        = "Eisuke Kita and Norio Kamiya"
+  author        = "Eisuke Kita and Norio Kamiya",
+  doi           = "10.1016/0965-9978(95)00067-4"
 }
 
 @Inproceedings{TrefftzSurvey,
