### 2022-02-02 #4112 @Nikoleta-v3
👋 @mscroggs &  @Balaje & @felipefr & @sigvaldm would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "NGSTrefftz: Add-on to NGSolve for Trefftz methods ". You can find more information at the top of this Github issue 🆙
