### 2022-02-06 #4112 @Nikoleta-v3
👋 @thelfer & @capitalaslash & @yangbai90 would any of you be willing to review this submission for JOSS?

The submission I'd like you to review is titled: "NGSTrefftz: Add-on to NGSolve for Trefftz methods ". You can find more information at the top of this Github issue 🆙
