### 2022-02-07 #4112 @thelfer
@Nikoleta-v3 Don't know Trefftz methods, but the method seems worth investigating. And it will be the occasion to learn a bit more about `NGSolve`, which I don't know either. If you don't expect the review for tomorrow,  I would be quite interested to review the paper. I could do that in the coming month.
