### 2022-02-07 #4135 @whedon
[REVIEW]: NGSTrefftz: Add-on to NGSolve for Trefftz methods
**Submitting author:** @PaulSt (<a href="http://orcid.org/0000-0001-5073-3366">Paul Stocker</a>)
**Repository:** https://github.com/PaulSt/NGSTrefftz
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.0
