### 2022-02-08 3c4621d
emb trefftz: allow for definedon-integrals
