### 2022-02-08 PaulSt/NGSTrefftz#2 @thelfer
Invalid url in the `README.md` file
This issue is created as part of the review of the JOSS paper (https://github.com/openjournals/joss-reviews/issues/4135).

The url to `NGSolver` is invalid in the `README.md` file
