### 2022-02-11 c10852a
Make test (#30)
* try make test on ubuntu

* fix tests

* test pypi pkg

* debut mac build

* better suggestions to find ngsolve

* replace ngstents with traversedim fix

* debug build ubuntu

* tmate session not working?

* remove old embTrefftz test using scipy.

* no more make test on os
