+++ .github/workflows/build.yml
@@ -36,11 +36,11 @@ jobs:
         run: |
              make -C $GITHUB_WORKSPACE/NGSTrefftz/make
              sudo make -C $GITHUB_WORKSPACE/NGSTrefftz/make install  
-      #- name: test NGSTrefftz
-      #  run: make -C $GITHUB_WORKSPACE/NGSTrefftz/make test
-      - name: Debugging with tmate
-        if: ${{ failure() }}
-        uses: mxschmitt/action-tmate@v3.6
+      - name: test NGSTrefftz
+        run: make -C $GITHUB_WORKSPACE/NGSTrefftz/make test
+      #- name: Debugging with tmate
+        #if: ${{ failure() }}
+        #uses: mxschmitt/action-tmate@v3
 
   macos:
     runs-on: macos-latest
@@ -76,7 +76,7 @@ jobs:
              sudo make -C $GITHUB_WORKSPACE/NGSTrefftz/make install  
 
 
-  publish:
+  pypi-publish:
     needs: [ubuntu,macos]
     runs-on: ubuntu-latest
     if: github.event_name == 'release' && github.event.action == 'created' #startsWith(github.ref, 'refs/tags/v') # && github.event_name == 'push' 
@@ -96,6 +96,20 @@ jobs:
           skip_existing: true
           user: __token__
           password: ${{ secrets.PYPI_API_TOKEN }}
+
+  pypi-test:
+    needs: [pypi-publish]
+    runs-on: ubuntu-latest
+    steps:
+      - name: install pkgs
+        run: |
+             sudo apt-get update && DEBIAN_FRONTEND="noninteractive" sudo apt-get -y install vim python3 python3-pip python3-distutils python3-tk libpython3-dev libxmu-dev tk-dev tcl-dev cmake git g++ libglu1-mesa-dev libblas-dev liblapack-dev
+             sudo apt-get install -y software-properties-common
+             sudo add-apt-repository universe
+             sudo add-apt-repository ppa:ngsolve/nightly -y
+             sudo apt-get install ngsolve -y
+      - name: install ngstrefftz
+        run: pip install ngstrefftz
 
 
   paper:
