+++ .github/tests/trefftz.py
@@ -7,6 +7,242 @@ from ngsolve import *
 from embt import *
 import time
 
+def DGwaveeqsys(fes,U0,v0,sig0,c,gD,fullsys=False, applyrhs = False,alpha=0.5,beta=0.5,gamma=1,mu=0,BB=1):
+    D = fes.mesh.dim - 1
+    # if not isinstance(BB,(int,float)): BB.spacedim = D
+    U = fes.TrialFunction()
+    V = fes.TestFunction()
+    gU = grad(U)
+    gV = grad(V)
+
+    v = gU[D]
+    sig = CoefficientFunction(tuple([-gU[i] for i in  range(D)]))*BB
+    w = gV[D]
+    tau = CoefficientFunction(tuple([-gV[i] for i in  range(D)]))*BB
+
+    vo = gU.Other()[D]
+    sigo = CoefficientFunction(tuple([-gU.Other()[i] for i in  range(D)]))*BB
+    wo = gV.Other()[D]
+    tauo = CoefficientFunction(tuple([-gV.Other()[i] for i in  range(D)]))*BB
+
+    h = specialcf.mesh_size
+    n = specialcf.normal(D+1)
+    n_t = n[D]/Norm(n)
+    n_x = CoefficientFunction( tuple([n[i]/Norm(n) for i in  range(D)]) )
+
+    mean_v = 0.5*(v+vo)
+    mean_w = 0.5*(w+wo)
+    mean_sig = 0.5*(sig+sigo)
+    mean_tau = 0.5*(tau+tauo)
+
+    jump_vx = ( v - vo ) * n_x
+    jump_wx = ( w - wo ) * n_x
+    jump_sigx = (( sig - sigo ) * n_x)
+    jump_taux = (( tau - tauo ) * n_x)
+
+    jump_vt = ( v - vo ) * n_t
+    jump_wt = ( w - wo ) * n_t
+    jump_sigt = ( sig - sigo ) * n_t
+    jump_taut = ( tau - tauo ) * n_t
+
+    jump_Ut = (U - U.Other()) * n_t
+    jump_Vt = (V - V.Other()) * n_t
+
+    timelike = n_x*n_x # n_t=0
+    spacelike = n_t**2 # n_x=0
+
+    delta=0.5
+    theta=1
+    gR=0
+
+    a = BilinearForm(fes)
+    if(fullsys==True):
+        HV = V.Operator("hesse")
+        a += SymbolicBFI( - v * (BB*sum([-HV[i*(D+2)] for i in range(D)]) + pow(c,-2)*HV[(D+1)*(D+1)-1]) )
+        if not isinstance(BB,(int,float)):
+            a += SymbolicBFI( - v * (BB.Diff(x)*(-gV[0]) + BB.Diff(y)*(-gV[1])) )
+        # a += SymbolicBFI(  sig*(CoefficientFunction([-HV[i,D] for i in range(D)]) + CoefficientFunction([HV[D,i] for i in range(D)])) )
+        HU = U.Operator("hesse")
+        a += SymbolicBFI(  mu * pow(c,2)
+                              * (-sum([HU[i*(D+2)] for i in range(D)]) + pow(c,-2)*HU[(D+1)*(D+1)-1])
+                              * (-sum([HV[i*(D+2)] for i in range(D)]) + pow(c,-2)*HV[(D+1)*(D+1)-1])
+                        )
+    #space like faces, w/o x jump ASSUME TENSOR MESH
+    if(applyrhs == False):
+        a += SymbolicBFI( spacelike * ( pow(c,-2)*IfPos(n_t,v,vo)*jump_wt + IfPos(n_t,sig,sigo)*jump_taut/BB ), VOL, skeleton=True )
[…]
