### 2022-02-11 2c3900b
add trafo and use CalcElMatAdd (with simd-catch) (#31)
* add trafo and use CalcElMatAdd (with simd-catch)

* add fix for PP->shape in unfitted case

* embtrefftz: CalcElMatAdd compatible with ExceptionNoSIMD
