+++ README.md
@@ -4,7 +4,7 @@
 [![PyPI](https://img.shields.io/pypi/v/ngstrefftz?color=blue&logo=pypi)](https://pypi.org/project/ngstrefftz/)
 [![GitHub Workflow Status](https://img.shields.io/github/workflow/status/PaulSt/NGSTrefftz/build?logo=github)](https://github.com/PaulSt/NGSTrefftz/actions)
 
-NGSTrefftz provides a framework to implement Trefftz finite element spaces for ![NGSolve](https://ngsolve.com), with several Trefftz spaces already implemented. Additionally, Trefftz-DG on tent-pitched meshes for the acoustic wave equation is implemented using meshes provided by ![ngstents](https://github.com/jayggg/ngstents). Furthermore, the package includes an implementation of the embedded Trefftz method.
+NGSTrefftz provides a framework to implement Trefftz finite element spaces for [NGSolve](https://www.ngsolve.com), with several Trefftz spaces already implemented. Additionally, Trefftz-DG on tent-pitched meshes for the acoustic wave equation is implemented using meshes provided by ![ngstents](https://github.com/jayggg/ngstents). Furthermore, the package includes an implementation of the embedded Trefftz method.
 
 ## Try it out!
 You can try out some jupyter notebooks:
@@ -18,7 +18,7 @@ docker run -p 8888:8888 ngstrefftz_jupyter
 ```
 
 ## Installing the package
-You need to have ![ngsolve](https://ngsolve.com/) installed. Then you can either:
+You need to have [NGSolve](https://www.ngsolve.com/) installed. Then you can either:
  * install using pip
 ```bash
 pip install ngstrefftz
