### 2022-02-11 PaulSt/NGSTrefftz#2 @PaulSt
Thank you @thelfer, should be fixed with 671deb29f49edf8fd944a76956dd537f328d4360
