### 2022-02-11 b38d293
Dockerpublish (#32)
* working on docker publish

* test run

* fix indent

* fix context?

* fix context try 2

* fix context try 3

* fix context try 4

* context??

* context???

* path context?

* file in context?

* context file

* context https?

* context????

* generate tags

* cleanup tags
