+++ .github/workflows/build.yml
@@ -111,6 +111,45 @@ jobs:
       - name: install ngstrefftz
         run: pip install ngstrefftz
 
+  docker:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v2
+      - name: Docker meta
+        id: meta
+        uses: docker/metadata-action@v3
+        with:
+          # list of Docker images to use as base name for tags
+          images: |
+            paulstdocker/ngstrefftz
+          # generate Docker tags based on the following events/attributes
+          tags: |
+            type=semver,pattern={{version}}
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v1
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v1
+      - name: Login to DockerHub
+        if: github.event_name == 'release' && github.event.action == 'created'
+        uses: docker/login-action@v1
+        with:
+          username: ${{ secrets.DOCKERHUB_USERNAME }}
+          password: ${{ secrets.DOCKERHUB_TOKEN }}
+      - name: Build and push
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          file: ".github/workflows/Dockerfile"
+          push: github.event_name == 'release' && github.event.action == 'created'
+          tags: ${{ steps.meta.outputs.tags }}
+          #labels: ${{ steps.meta.outputs.labels }}
+      - name: test docker
+        uses: docker/build-push-action@v2
+        with:
+          context: .
+          file: ".github/workflows/Dockerfile"
+          target: test
 
   paper:
     runs-on: ubuntu-latest
