### 2022-02-11 e1a98dc
update requirements on docker workflow
