+++ .github/workflows/build.yml
@@ -112,6 +112,7 @@ jobs:
         run: pip install ngstrefftz
 
   docker:
+    needs: [ubuntu,macos]
     runs-on: ubuntu-latest
     steps:
       - name: Checkout
@@ -137,11 +138,12 @@ jobs:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
       - name: Build and push
+        if: github.event_name == 'release' && github.event.action == 'created'
         uses: docker/build-push-action@v2
         with:
           context: .
           file: ".github/workflows/Dockerfile"
-          push: github.event_name == 'release' && github.event.action == 'created'
+          push: true
           tags: ${{ steps.meta.outputs.tags }}
           #labels: ${{ steps.meta.outputs.labels }}
       - name: test docker
