### 2022-02-11 1e95677
remove target test in dockerbuild
