+++ .github/workflows/build.yml
@@ -114,6 +114,7 @@ jobs:
   docker:
     needs: [ubuntu,macos]
     runs-on: ubuntu-latest
+    if: github.event_name == 'release' && github.event.action == 'created'
     steps:
       - name: Checkout
         uses: actions/checkout@v2
@@ -132,13 +133,11 @@ jobs:
       - name: Set up Docker Buildx
         uses: docker/setup-buildx-action@v1
       - name: Login to DockerHub
-        if: github.event_name == 'release' && github.event.action == 'created'
         uses: docker/login-action@v1
         with:
           username: ${{ secrets.DOCKERHUB_USERNAME }}
           password: ${{ secrets.DOCKERHUB_TOKEN }}
       - name: Build and push
-        if: github.event_name == 'release' && github.event.action == 'created'
         uses: docker/build-push-action@v2
         with:
           context: .
@@ -146,12 +145,6 @@ jobs:
           push: true
           tags: ${{ steps.meta.outputs.tags }}
           #labels: ${{ steps.meta.outputs.labels }}
-      - name: test docker
-        uses: docker/build-push-action@v2
-        with:
-          context: .
-          file: ".github/workflows/Dockerfile"
-          target: test
 
   paper:
     runs-on: ubuntu-latest
