### 2022-02-12 4b28993
Mixed (#34)
* add trafo and use CalcElMatAdd (with simd-catch)

* add fix for PP->shape in unfitted case

* embtrefftz: CalcElMatAdd compatible with ExceptionNoSIMD

* Starting with trial/test-FEs in op assembly

* mixed working?

* add test

* fix test and update defaults args

* fix mixed poisson problem and add test

Co-authored-by: Christoph Lehrenfeld <christoph.lehrenfeld@gmail.com>
Co-authored-by: Fabian Heimann <Fabian.Heimann@stud.uni-goettingen.de>
