+++ src/embtrefftz.cpp
@@ -11,7 +11,7 @@ namespace ngbla
   {
     static Timer t ("LapackSVD");
     RegionTimer reg (t);
-    ngbla::integer m = A.Width (), n = A.Height ();
+    ngbla::integer n = A.Width (), m = A.Height ();
     Vector<> S (min (n, m));
     Array<double> work (4 * m * m + 6 * m + m + 100);
     Array<int> iwork (max (n, m) * 9);
@@ -36,7 +36,7 @@ namespace ngbla
   {
     static Timer t ("LapackSVD");
     RegionTimer reg (t);
-    ngbla::integer m = A.Width (), n = A.Height ();
+    ngbla::integer n = A.Width (), m = A.Height ();
     Vector<> S (min (n, m));
     Array<Complex> work (4 * m * m + 6 * m + m + 100);
     Array<int> iwork (max (n, m) * 9);
@@ -72,7 +72,9 @@ namespace ngbla
     CalcSVD (AA, U, V);
 #endif
     A = 0.0;
-    A.Diag (0) = AA.Diag ();
+    // A.Diag(0)=AA.Diag();
+    for (int i = 0; i < min (A.Width (), A.Height ()); i++)
+      A (i, i) = AA (i, i);
   }
 
   template void
@@ -90,11 +92,19 @@ namespace ngcomp
   template <class SCAL>
   std::tuple<shared_ptr<BaseMatrix>, shared_ptr<BaseVector>>
   EmbTrefftz (shared_ptr<SumOfIntegrals> bf, shared_ptr<FESpace> fes,
-              double eps, shared_ptr<SumOfIntegrals> lf)
+              shared_ptr<SumOfIntegrals> lf, double eps,
+              shared_ptr<FESpace> test_fes, int tndof)
   {
     static Timer svdtt ("svdtrefftz");
     RegionTimer reg (svdtt);
     LocalHeap lh (1000 * 1000 * 1000);
+
+    bool mixed_mode = true;
+    if (test_fes == nullptr)
+      {
+        mixed_mode = false;
+        test_fes = fes;
+      }
 
     auto ma = fes->GetMeshAccess ();
 
@@ -121,6 +131,8 @@ namespace ngcomp
 
     ma->IterateElements (VOL, lh, [&] (auto ei, LocalHeap &mlh) {
       HeapReset hr (mlh);
+      Array<DofId> test_dofs;
+      test_fes->GetDofNrs (ei, test_dofs);
       Array<DofId> dofs;
       fes->GetDofNrs (ei, dofs);
 
@@ -136,9 +148,12 @@ namespace ngcomp
         return; // escape lambda
 
       auto &trafo = ma->GetTrafo (ei, mlh);
-      auto &fel = fes->GetFE (ei, mlh);
 
-      FlatMatrix<SCAL> elmat (dofs.Size (), mlh);
+      auto &test_fel = test_fes->GetFE (ei, mlh);
+      auto &trial_fel = fes->GetFE (ei, mlh);
+
+      FlatMatrix<SCAL> elmat (test_dofs.Size (), dofs.Size (), mlh);
+
       elmat = 0.0;
       bool symmetric_so_far = true;
 
@@ -153,8 +168,18 @@ namespace ngcomp
                   = trafo.AddDeformation (bfi->GetDeformation ().get (), mlh);
               try
                 {
-                  bfi->CalcElementMatrixAdd (fel, mapped_trafo, elmat,
-                                             symmetric_so_far, mlh);
+                  if (mixed_mode)
+                    {
+                      const auto &mixed_fel
+                          = MixedFiniteElement (trial_fel, test_fel);
+                      bfi->CalcElementMatrixAdd (mixed_fel, mapped_trafo,
+                                                 elmat, symmetric_so_far, mlh);
[…]
