+++ src/embtrefftz.hpp
@@ -10,7 +10,8 @@ namespace ngcomp
   template <class SCAL>
   std::tuple<shared_ptr<BaseMatrix>, shared_ptr<BaseVector>>
   EmbTrefftz (shared_ptr<SumOfIntegrals> bf, shared_ptr<FESpace> fes,
-              double eps = 10e-10, shared_ptr<SumOfIntegrals> lf = nullptr);
+              shared_ptr<SumOfIntegrals> lf, double eps,
+              shared_ptr<FESpace> fes_test, int tndof);
 }
 
 #ifdef NGS_PYTHON
