### 2022-02-15 PaulSt/NGSTrefftz#3 @thelfer
Hi @PaulSt

I updated the submodules and went a bit further:

```bash
In file included from /home/th202608/codes/NGSTrefftz/master/src/NGSTrefftz/src/trefftzfespace.hpp:5,
                 from /home/th202608/codes/NGSTrefftz/master/src/NGSTrefftz/src/python_trefftz.cpp:7:
/home/th202608/codes/NGSTrefftz/master/src/NGSTrefftz/src/planewavefe.hpp: In constructor ‘ngfem::PlaneWaveElement<D>::PlaneWaveElement(int, int, ngfem::ELEMENT_TYPE, ngbla::Vec<S>, double, float, int)’:
/home/th202608/codes/NGSTrefftz/master/src/NGSTrefftz/src/planewavefe.hpp:23:56: error: cannot deduce template arguments for ‘Matrix’ from ()
             ScalarMappedElement<D>(andof, aord, Matrix(), aeltype, aelcenter, aelsize, ac),
```
