+++ src/planewavefe.hpp
@@ -20,7 +20,7 @@ namespace ngfem
 
         public:
             PlaneWaveElement(int andof, int aord, ELEMENT_TYPE aeltype, Vec<D> aelcenter = 0, double aelsize = 1, float ac = 1.0, int aconj=1) :
-            ScalarMappedElement<D>(andof, aord, Matrix(), aeltype, aelcenter, aelsize, ac),
+            ScalarMappedElement<D>(andof, aord, Matrix<>(), aeltype, aelcenter, aelsize, ac),
             conj(aconj)
             {;}
 
