### 2022-02-15 PaulSt/NGSTrefftz#3 @PaulSt
Ok, great that this worked!
Unfortunately, I am unable to reproduce the error you are getting now.
I tested with gcc/g++ 11.1.0 and clang 13.0.0, what are you using?
I tried a possible fix with the newest commit f0995226bc7dad86fce54ae633163d8e1da62069 hope it works!
