### 2022-02-15 PaulSt/NGSTrefftz#3 @thelfer
@PaulSt I am using gcc-8.3.0 on debian buster
