### 2022-02-15 PaulSt/NGSTrefftz#3 @thelfer
With you fix, it goes a bit further. Now I have a huge log. Please find the attached file

[out.zip](https://github.com/PaulSt/NGSTrefftz/files/8074203/out.zip)
.
