### 2022-02-15 PaulSt/NGSTrefftz#3 @thelfer
BTW, it indeed builds with `gcc-11.0`
