### 2022-02-15 PaulSt/NGSTrefftz#4 @thelfer
Build directory must have a specific location to run the tests
This issue is created as part of the review of the JOSS paper (openjournals/joss-reviews#4135).

It seems that running the tests requires to build `NGSTrefftz` in a subdirectory of the source directory, as indicated in the `README` file.

Otherwise, this is seems a terrible practice and is probably incompatible with most package manager, hence the issue.
