### 2022-02-17 3c2bc6c
copy cmake tests to build dir
