### 2022-02-17 cdd0e97
copy cmake tests to build dir
