### 2022-02-17 6fa8ba8
update dependencies in readme
