### 2022-02-17 PaulSt/NGSTrefftz#3 @PaulSt
Thank you again, I am happy it compiles now! I fixed the error from
> [out.zip](https://github.com/PaulSt/NGSTrefftz/files/8074203/out.zip)

and it should compile with gcc-8 now. However, for the future I would suggest clang >= 10 or at least gcc >= 9, I updated the readme with the dependencies for the build.
