### 2022-02-17 PaulSt/NGSTrefftz#5 @thelfer
Structure of the installation directory
This issue is created as part of the review of the JOSS paper (https://github.com/openjournals/joss-reviews/issues/4135).

The installation directory of `NGSTrefftz` is currently as follows:

~~~~
$ tree -d .
.
├── ngstents
└── ngstrefftz
~~~~

This structure is not standard and worse, not compatible with the one of `NGSolve` which installs its `python` module in `lib/python3/dist-packages/`.
