### 2022-02-17 PaulSt/NGSTrefftz#6 @thelfer
Tests does not work out of the box
This issue is created as part of the review of the JOSS paper (https://github.com/openjournals/joss-reviews/issues/4135).

The tests do not work out of the box. To run them, I had to:

1. install the package
2. update my environment variables (`PYTHONPATH` in practice)
3. run the tests with `make test`

This is fairly cumbersome and unsual and incompatible with most common practices and package managers.
