### 2022-02-17 PaulSt/NGSTrefftz#6 @thelfer
Note that requiring `ngsolve` to be correctly installed and accessible (thanks to environment variables like `PYTHONPATH` and `LD_LIBRARY_PATH`) is just fine IMHO.
