### 2022-02-18 PaulSt/NGSTrefftz#5 @PaulSt
Hmm, are you passing `DCMAKE_INSTALL_PREFIX` to cmake?

I believe, you will get the behavior you are looking for when you do *not* tell cmake a install prefix. Then, the install directory of ngsolve is used, which should then be `lib/python3/dist-packages/`

This directory should also be in the `PYTHONPATH` (for ngsolve to work) so that the tests should work out of the box https://github.com/PaulSt/NGSTrefftz/issues/6

If I am understanding correctly, and this indeed solves the issue, then I could put some kind of warning into the cmake in the case of `DCMAKE_INSTALL_PREFIX` to inform the user that the path deviates from the ngsolve install path and that this needs to be added to the pythonpath.
