### 2022-02-18 PaulSt/NGSTrefftz#5 @thelfer
Concerning the tests, you seem to suggest that one have to install `NGSTrefftz` **before** running the tests. This is also non standard. The standard way is:

```bash
$ cmake
$ make
$ make test
$ make install
```
