### 2022-02-19 ad9e06a
improve cmake, tests and install dir (#35)
* improve cmake, tests and install dir

* install needs sudo
