+++ src/python_trefftz.cpp
@@ -8,12 +8,9 @@
 #include "specialcoefficientfunction.hpp"
 #include "twavetents.hpp"
 #include "embtrefftz.hpp"
-
-#ifdef EXT_TTEST
 #include "mesh1dtents.hpp"
 #include "monomialfespace.hpp"
 //#include "airy.cpp"
-#endif
 
 
 PYBIND11_MODULE(_trefftz,m) {
@@ -25,11 +22,9 @@ PYBIND11_MODULE(_trefftz,m) {
     ExportSpecialCoefficientFunction(m);
     ExportTWaveTents(m);
     ExportEmbTrefftz(m);
-#ifdef EXT_TTEST
     ExportMesh1dTents(m);
     ExportMonomialFESpace(m);
     //ExportStdMathFunction<GenericAiry>(m, "airy", "airy function");
     //ExportStdMathFunction<GenericAiryP>(m, "airyp", "airyp function");
-#endif
 }
 
