+++ test/embt.py
@@ -1,3 +1,5 @@
+import sys, os
+sys.path.append(os.path.join(os.path.dirname(sys.path[0])))
 from ngsolve import *
 from ngstrefftz import *
 import time
