+++ src/CMakeLists.txt
@@ -1,30 +1,40 @@
+########################################################
 project(ngstrefftz)
+########################################################
+
+option(USE_LAPACK "build ngstrefftz with LAPACK" ON)
+
+### Requirements
+file(TO_CMAKE_PATH "${PROJECT_BINARY_DIR}/CMakeLists.txt" LOC_PATH)
+if(EXISTS "${LOC_PATH}")
+    message(FATAL_ERROR "You cannot build in a source directory (or any directory with a CMakeLists.txt file). Please make a build subdirectory. Feel free to remove CMakeCache.txt and CMakeFiles.")
+endif()
 cmake_minimum_required(VERSION 3.1)
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
-
-option(EXT_TTEST "Add helpful functions for testing" OFF)
-option(USE_LAPACK "build ngstrefftz with LAPACK" ON)
-
-
 find_package(NGSolve CONFIG REQUIRED
   HINTS $ENV{NETGENDIR}/.. $ENV{NETGENDIR}/../Resources/CMake /opt/netgen/ /Applications/Netgen.app/Contents/Resources/CMake C:/netgen
     )
-# check if CMAKE_INSTALL_PREFIX is set by user, if not install in NGSolve python dir
+
+### check if CMAKE_INSTALL_PREFIX is set by user, if not install in NGSolve python dir
 if(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
-    set(CMAKE_INSTALL_PREFIX ${NGSOLVE_INSTALL_DIR}/${NGSOLVE_INSTALL_DIR_PYTHON} CACHE PATH "Install dir" FORCE)
+    set(CMAKE_INSTALL_PREFIX ${NGSOLVE_INSTALL_DIR} CACHE PATH "Install dir" FORCE)
 endif(CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT)
+set(NGSTREFFTZ_INSTALL_DIR_PYTHON ${CMAKE_INSTALL_PREFIX}/${NGSOLVE_INSTALL_DIR_PYTHON}/ngstrefftz)
 
-
+### add ngstents dependency if needed
 if (NOT TARGET _pytents)
-include_directories(${CMAKE_CURRENT_SOURCE_DIR}/../external_dependencies/ngstents/src)
-add_ngsolve_python_module(_pytents
-  ${CMAKE_CURRENT_SOURCE_DIR}/../external_dependencies/ngstents/src/python_tents.cpp
-  ${CMAKE_CURRENT_SOURCE_DIR}/../external_dependencies/ngstents/src/tents.cpp
-  )
-install(TARGETS _pytents DESTINATION ngstents)
+    include_directories(${CMAKE_CURRENT_SOURCE_DIR}/../external_dependencies/ngstents/src)
+    set(NGSTENTS_INSTALL_DIR_PYTHON ${CMAKE_INSTALL_PREFIX}/${NGSOLVE_INSTALL_DIR_PYTHON}/ngstents)
+    add_ngsolve_python_module(_pytents
+      ${CMAKE_CURRENT_SOURCE_DIR}/../external_dependencies/ngstents/src/python_tents.cpp
+      ${CMAKE_CURRENT_SOURCE_DIR}/../external_dependencies/ngstents/src/tents.cpp
+      )
+    set_target_properties(_pytents PROPERTIES LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/ngstents")
[…]
