### 2022-02-19 d99938d
improve cmake, tests and install dir (#35)
* improve cmake, tests and install dir

* install needs sudo
