### 2022-02-19 47f6443
Clean tfes (#36)
* trefftz basis with loops instead of recursion

* cleanup fespace
