+++ src/trefftzfespace.cpp
@@ -295,6 +295,44 @@ namespace ngcomp
     static RegisterFESpace<TrefftzFESpace> initi_trefftz ("trefftzfespace");
 
 
+
+
+    //////////////////////////// Trefftz basis ////////////////////////////
+
+    template<int D,typename TFunc>
+    void TraversePol( int order, const TFunc &func)
+    {
+        switch(D)
+        {
+            case 0:
+                break;
+            case 1:
+                for (int i = 0, ii = 0; i <= order; i++)
+                                func(ii++,Vec<1,int>{i});
+                break;
+            case 2:
+                for (int i = 0, ii = 0; i <= order; i++)
+                    for (int j = 0; j <= order-i; j++)
+                                func(ii++,Vec<2,int>{j,i});
+                break;
+            case 3:
+                for (int i = 0, ii = 0; i <= order; i++)
+                    for (int j = 0; j <= order-i; j++)
+                        for (int k = 0; k <= order-i-j; k++)
+                                func(ii++,Vec<3,int>{k,j,i});
+                break;
+            case 4:
+                for (int i = 0, ii = 0; i <= order; i++)
+                    for (int j = 0; j <= order-i; j++)
+                        for (int k = 0; k <= order-i-j; k++)
+                            for (int l = 0; l <= order-i-j-k; l++)
+                                func(ii++,Vec<4,int>{l,k,j,i});
+                break;
+            default:
+                throw Exception("TraverseDimensions: too many dimensions!");
+        }
+    }
+
     // k-th coeff of Legendre polynomial of degree n in monomial basis
     constexpr double LegCoeffMonBasis(int n, int k)
     {
@@ -316,7 +354,10 @@ namespace ngcomp
         return coeff;
     }
 
-
+    constexpr int factorial(int n)
+    {
+        return n>1 ? n * factorial(n-1) : 1;
+    }
 
 
 
@@ -328,38 +369,13 @@ namespace ngcomp
         const int npoly = (BinCoeff(D+1 + ord, ord));
         Matrix<> trefftzbasis(ndof,npoly);
         trefftzbasis = 0;
-        Vec<D+1, int>  coeff = 0;
-        for(int b=0;b<ndof;b++)
+        for(int basis=0;basis<ndof;basis++)
         {
             int tracker = 0;
-            TB_inner(ord, trefftzbasis, coeff, b, D+1, tracker, basistype);
-        }
-        MatToCSR(trefftzbasis.Rows(fowave,ndof),tb);
-        return tb;
-    }
-
-
-    template<int D>
-    void TWaveBasis<D> :: TB_inner(int ord, Matrix<> &trefftzbasis, Vec<D+1, int> coeffnum, int basis, int dim, int &tracker, int basistype, double wavespeed)
-    {
-        if (dim>0)
-        {
-            while(coeffnum(dim-1)<=ord)
-            {
-                TB_inner(ord,trefftzbasis,coeffnum,basis, dim-1, tracker, basistype, wavespeed);
-                coeffnum(dim-1)++;
-            }
-        }
-        else
-        {
-            int sum=0;
-            for(int i=0;i<D+1;i++)
-                sum += coeffnum(i);
-            if(sum<=ord)
-            {
+            TraversePol<D+1>( ord, [&](int i, Vec<D+1,int> coeff){
                 if(tracker >= 0) tracker++;
-                int indexmap = PolBasis::IndexMap2<D>(coeffnum, ord);
-                int k = coeffnum(D);
+                int indexmap = PolBasis::IndexMap2<D>(coeff, ord);
+                int k = coeff(D);
[…]
