### 2022-02-19 806defa
Clean tfes (#36)
* trefftz basis with loops instead of recursion

* cleanup fespace
