+++ src/trefftzfespace.cpp
@@ -375,6 +375,42 @@ namespace ngcomp
 
   static RegisterFESpace<TrefftzFESpace> initi_trefftz ("trefftzfespace");
 
+  //////////////////////////// Trefftz basis ////////////////////////////
+
+  template <int D, typename TFunc>
+  void TraversePol (int order, const TFunc &func)
+  {
+    switch (D)
+      {
+      case 0:
+        break;
+      case 1:
+        for (int i = 0, ii = 0; i <= order; i++)
+          func (ii++, Vec<1, int>{ i });
+        break;
+      case 2:
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            func (ii++, Vec<2, int>{ j, i });
+        break;
+      case 3:
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
+              func (ii++, Vec<3, int>{ k, j, i });
+        break;
+      case 4:
+        for (int i = 0, ii = 0; i <= order; i++)
+          for (int j = 0; j <= order - i; j++)
+            for (int k = 0; k <= order - i - j; k++)
+              for (int l = 0; l <= order - i - j - k; l++)
+                func (ii++, Vec<4, int>{ l, k, j, i });
+        break;
+      default:
+        throw Exception ("TraverseDimensions: too many dimensions!");
+      }
+  }
+
   // k-th coeff of Legendre polynomial of degree n in monomial basis
   constexpr double LegCoeffMonBasis (int n, int k)
   {
@@ -406,6 +442,8 @@ namespace ngcomp
     return coeff;
   }
 
+  constexpr int factorial (int n) { return n > 1 ? n * factorial (n - 1) : 1; }
+
   template <int D>
   CSR TWaveBasis<D>::Basis (int ord, int basistype, int fowave)
   {
@@ -415,101 +453,185 @@ namespace ngcomp
     const int npoly = (BinCoeff (D + 1 + ord, ord));
     Matrix<> trefftzbasis (ndof, npoly);
     trefftzbasis = 0;
-    Vec<D + 1, int> coeff = 0;
-    for (int b = 0; b < ndof; b++)
+    for (int basis = 0; basis < ndof; basis++)
       {
         int tracker = 0;
-        TB_inner (ord, trefftzbasis, coeff, b, D + 1, tracker, basistype);
+        TraversePol<D + 1> (ord, [&] (int i, Vec<D + 1, int> coeff) {
+          if (tracker >= 0)
+            tracker++;
+          int indexmap = PolBasis::IndexMap2<D> (coeff, ord);
+          int k = coeff (D);
+          if (k == 0 || k == 1)
+            {
+              switch (basistype)
+                {
+                case 0:
+                  if (tracker > basis)
+                    {
+                      // trefftzbasis( i, setbasis++ ) = 1.0; //set the l-th
+                      // coeff to 1
+                      trefftzbasis (basis, indexmap) = 1;
+                      tracker = -1;
+                    }
+                  // i += ndof-1;	//jump to time = 2 if i=0
+                  break;
+                case 1:
+                  if ((k == 0 && basis < BinCoeff (D + ord, ord))
+                      || (k == 1 && basis >= BinCoeff (D + ord, ord)))
+                    {
+                      trefftzbasis (basis, indexmap) = 1;
+                      for (int exponent : coeff.Range (0, D))
+                        trefftzbasis (basis, indexmap)
+                            *= LegCoeffMonBasis (basis, exponent);
+                    }
+                  break;
+                case 2:
+                  if ((k == 0 && basis < BinCoeff (D + ord, ord))
[…]
