### 2022-02-19 e91afa0
cmake add debug compiler options
