### 2022-02-19 PaulSt/NGSTrefftz#5 @PaulSt
Thank you for clearing that up, this was very helpful! I updated the cmake file with ad9e06a33e4319a9f76250d7c008fd25c5beb8b3 to hopefully meet all specifications!
