### 2022-02-20 7c3dcb7
Better cmake (#37)
* improve cmake, tests and install dir

* install needs sudo

* fix pip path error

* update pypi workflow

* fix version

* rm version.py
