### 2022-02-21 #4135 @thelfer
I currently have a working version of `NGSTrefftz`, after resolution of various issues. Review starts now !
