### 2022-02-22 9b9f863
cmake realease/debug options
