+++ src/embtrefftz.cpp
@@ -99,6 +99,9 @@ namespace ngcomp
     RegionTimer reg (svdtt);
     LocalHeap lh (1000 * 1000 * 1000);
 
+    if (eps == 0 && tndof == 0 && test_fes == nullptr)
+      throw Exception ("Need to specify eps, tndof, or test_fes");
+
     bool mixed_mode = true;
     if (test_fes == nullptr)
       {
@@ -153,10 +156,8 @@ namespace ngcomp
       auto &trial_fel = fes->GetFE (ei, mlh);
 
       FlatMatrix<SCAL> elmat (test_dofs.Size (), dofs.Size (), mlh);
-
       elmat = 0.0;
       bool symmetric_so_far = true;
-
       int bfi_ind = 0;
       while (bfi_ind < bfis[VOL].Size ())
         {
@@ -198,14 +199,15 @@ namespace ngcomp
       // assumption here: all (active) elements have the same number of (weak)
       // Trefftz fcts.
       int nz = 0;
-      if (mixed_mode)
-        nz = trial_fel.GetNDof () - test_fel.GetNDof ();
-      else if (tndof)
+      if (tndof)
         nz = tndof;
       else
-        for (auto sv : elmat.Diag ())
-          if (abs (sv) < eps)
-            nz++;
+        {
+          nz = trial_fel.GetNDof () - test_fel.GetNDof ();
+          for (int i = 0; i < min (elmat.Width (), elmat.Height ()); i++)
+            if (abs (elmat (i, i)) < eps)
+              nz++;
+        }
 
       std::call_once (init_flag, [&] () {
         TableCreator<int> creator (ma->GetNE (VOL));
@@ -310,7 +312,7 @@ void ExportEmbTrefftz (py::module m)
 
         return ngcomp::EmbTrefftz<double> (bf, fes, lf, eps, test_fes, tndof);
       },
-      py::arg ("bf"), py::arg ("fes"), py::arg ("lf"), py::arg ("eps") = 1e-12,
+      py::arg ("bf"), py::arg ("fes"), py::arg ("lf"), py::arg ("eps") = 0,
       py::arg ("test_fes") = nullptr, py::arg ("tndof") = 0);
 
   m.def (
@@ -326,7 +328,7 @@ void ExportEmbTrefftz (py::module m)
         return std::get<0> (ngcomp::EmbTrefftz<double> (bf, fes, nullptr, eps,
                                                         test_fes, tndof));
       },
-      py::arg ("bf"), py::arg ("fes"), py::arg ("eps") = 1e-12,
+      py::arg ("bf"), py::arg ("fes"), py::arg ("eps") = 0,
       py::arg ("test_fes") = nullptr, py::arg ("tndof") = 0);
 }
 #endif // NGS_PYTHON
