### 2022-02-22 956cf4a
fix mixed with l.d. constraints
