### 2022-02-22 ad92a65
remove default debug if git in cmake
