### 2022-03-01 PaulSt/NGSTrefftz#7 @thelfer
Missing community guidelines
This issue is created as part of the review of the JOSS paper (openjournals/joss-reviews#4135).

The review implies checking if proper community guidelines are present.  Here are the requirements:

> Community guidelines: Are there clear guidelines for third parties wishing to
>
>  1) Contribute to the software
>  2) Report issues or problems with the software
>  3) Seek support

I did not find such guidelines. You may want to have a look at this example from one of my personal projects: <https://github.com/thelfer/MFrontGenericInterfaceSupport/blob/master/CONTRIBUTING.md>
