### 2022-03-01 PaulSt/NGSTrefftz#8 @thelfer
API documentation
This issue is created as part of the review of the JOSS paper (https://github.com/openjournals/joss-reviews/issues/4135).

JOSS review criteria imply judging the API documentation (<https://joss.readthedocs.io/en/latest/review_criteria.html>).

If I am not mistaken, I only saw tutorials, but no pages describing the API.
