### 2022-03-03 f7ab482
add docstrings and sphinx-quickstart
