+++ src/twavetents.cpp
@@ -12,7 +12,6 @@ namespace ngcomp
         return (T(0) <= val) - (val < T(0));
     }
 
-    TrefftzTents::~TrefftzTents() {};
 
     template<int D>
     inline void  TWaveTents<D> :: Solve(FlatMatrix<double> a, FlatVector<double> b)
@@ -939,10 +938,7 @@ void DeclareETClass(py::module &m, std::string typestr)
     std::string pyclass_name = typestr;
     py::class_<PyETclass, shared_ptr<PyETclass>, TrefftzTents>(m, pyclass_name.c_str())
         //.def(py::init<>())
-        .def("Propagate", &PyETclass::Propagate)
         .def("MakeWavefront", &PyETclass::MakeWavefront)
-        .def("SetInitial", &PyETclass::SetInitial)
-        .def("SetBoundaryCF", &PyETclass::SetBoundaryCF)
         .def("GetWavefront", &PyETclass::GetWavefront)
         .def("Error", &PyETclass::Error)
         .def("L2Error", &PyETclass::L2Error)
@@ -956,7 +952,10 @@ void DeclareETClass(py::module &m, std::string typestr)
 
 void ExportTWaveTents(py::module m)
 {
-    py::class_<TrefftzTents, shared_ptr<TrefftzTents>>(m, "TrefftzTents");
+    py::class_<TrefftzTents, shared_ptr<TrefftzTents>>(m, "TrefftzTents")
+        .def("Propagate", &TrefftzTents::Propagate, "Solve tent slab")
+        .def("SetInitial", &TrefftzTents::SetInitial, "Set initial condition")
+        .def("SetBoundaryCF", &TrefftzTents::SetBoundaryCF, "Set boundary condition");
 
     DeclareETClass<TWaveTents<1>, 1>(m, "TWaveTents1");
     DeclareETClass<TWaveTents<2>, 2>(m, "TWaveTents2");
@@ -983,7 +982,14 @@ void ExportTWaveTents(py::module m)
                       tr = make_shared<QTWaveTents<2>>(order,tps,wavespeedcf,BBcf);
               }
               return tr;
-          }, "Create Wavetent object",
+          }, R"mydelimiter(
+                Create solver for acoustiv wave equation on tent-pitched mesh.
+
+                :param order: Polynomial order of the Trefftz space.
+                :param tps: Tent-pitched slab.
+                :param wavespeedcf: PDE Coefficient
+                :param BB: PDE Coefficient
+            )mydelimiter",
         py::arg("order"), py::arg("tps"), py::arg("wavespeedcf"), py::arg("BBcf")=nullptr
             );
 
