+++ src/twavetents.hpp
@@ -11,7 +11,12 @@ namespace ngcomp
         public:
             TrefftzTents(){;}
             virtual int dimensio(){return 0;}
-            virtual ~TrefftzTents() = 0;
+            virtual void Propagate() 
+                { throw Exception("TrefftzTents virtual!"); }
+            virtual void SetInitial(shared_ptr<CoefficientFunction> init) 
+                { throw Exception("TrefftzTents virtual!"); }
+            virtual void SetBoundaryCF(shared_ptr<CoefficientFunction> abddatum) 
+                { throw Exception("TrefftzTents virtual!"); }
     };
 
     template<int D>
@@ -85,13 +90,13 @@ namespace ngcomp
                 }
             }
 
-            void Propagate();
+            void Propagate() override;
 
             Matrix<> MakeWavefront( shared_ptr<CoefficientFunction> cf, double time = 0);
 
             Matrix<> GetWavefront() {return wavefront;}
 
-            void SetInitial(shared_ptr<CoefficientFunction> init) {
+            void SetInitial(shared_ptr<CoefficientFunction> init) override {
                 wavefront = MakeWavefront(init);
                 if(init->Dimension()==D+1){
                     fosystem=1;
@@ -99,7 +104,7 @@ namespace ngcomp
                 }
             }
 
-            void SetBoundaryCF(shared_ptr<CoefficientFunction> abddatum) { bddatum = abddatum;}
+            void SetBoundaryCF(shared_ptr<CoefficientFunction> abddatum) override { bddatum = abddatum;}
 
             double Error(Matrix<> wavefront, Matrix<> wavefront_corr);
 
@@ -166,7 +171,6 @@ namespace ngcomp
             }
 
             void Propagate();
-
     };
 
 }
