### 2022-03-03 8c98fc1
add docstrings and sphinx-quickstart
