+++ doc/conf.py
@@ -1,0 +1,133 @@
+# Configuration file for the Sphinx documentation builder.
+#
+# This file only contains a selection of the most common options. For a full
+# list see the documentation:
+# https://www.sphinx-doc.org/en/master/usage/configuration.html
+
+# -- Path setup --------------------------------------------------------------
+
+# If extensions (or modules to document with autodoc) are in another directory,
+# add these directories to sys.path here. If the directory is relative to the
+# documentation root, use os.path.abspath to make it absolute, like shown here.
+#
+# import os
+# import sys
+# sys.path.insert(0, os.path.abspath('.'))
+
+
+# -- Project information -----------------------------------------------------
+
+project = 'NGSTrefftz'
+copyright = '2022, Paul Stocker'
+author = 'Paul Stocker'
+
+# The full version, including alpha/beta/rc tags
+release = 'v0.0.4'
+
+
+# -- General configuration ---------------------------------------------------
+
+# Add any Sphinx extension module names here, as strings. They can be
+# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
+# ones.
+# extensions = [ "myst_nb" ]
+extensions = ["sphinx.ext.autodoc","sphinx.ext.mathjax","sphinx.ext.todo",
+              "IPython.sphinxext.ipython_console_highlighting", "IPython.sphinxext.ipython_directive",
+              "nbsphinx",
+              "m2r2"
+              ]
+# source_suffix = ['.rst', '.md']
+
+# START nbsphinx stuff
+#increase timeout for cell execution, since some files take long to execute
+nbsphinx_timeout = 100000
+
+# If True, the build process is continued even if an exception occurs:
+nbsphinx_allow_errors = False
+
+# This is processed by Jinja2 and inserted before each notebook
+nbsphinx_prolog = r"""
+.. raw:: html
+
+    <style>
+        .p-Widget {
+            height: 400px;
+        }
+        .dg.main {
+            margin-left: 0px;
+        }
+        div.p-Widget div div div div.dg ul li {
+            list-style: none;
+            margin-left: 0px;
+        }
+        div.p-Widget div div div div.dg ul li div.dg {
+            margin-bottom: 0px;
+        }
+    </style>
+
+.. only:: html
+    .. role:: raw-html(raw)
+        :format: html
+
+"""
+
+# END nbsphinx stuff
+
+# The name of the Pygments (syntax highlighting) style to use.
+pygments_style = 'sphinx'
+
+# If true, `todo` and `todoList` produce output, else they produce nothing.
+todo_include_todos = False
+
+# Add any paths that contain templates here, relative to this directory.
+templates_path = ['_templates']
+
+# List of patterns, relative to source directory, that match files and
+# directories to ignore when looking for source files.
+# This pattern also affects html_static_path and html_extra_path.
+exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'paper']
+
+# LaTeX
+latex_engine = 'xelatex'
+latex_elements = {
+    'fontpkg': r'''
+\setmainfont{DejaVu Serif}
+\setsansfont{DejaVu Sans}
+\setmonofont{DejaVu Sans Mono}
+''',
+    'preamble': r'''
+\usepackage[titles]{tocloft}
+\cftsetpnumwidth {1.25cm}\cftsetrmarg{1.5cm}
+\setlength{\cftchapnumwidth}{0.75cm}
+\setlength{\cftsecindent}{\cftchapnumwidth}
+\setlength{\cftsecnumwidth}{1.25cm}
+''',
+    'fncychap': r'\usepackage[Bjornstrup]{fncychap}',
+    'printindex': r'\footnotesize\raggedright\printindex',
[…]
