+++ docs/requirements.txt
@@ -1,0 +1,79 @@
+alabaster==0.7.12
+argon2-cffi==21.3.0
+argon2-cffi-bindings==21.2.0
+asttokens==2.0.5
+attrs==21.4.0
+Babel==2.9.1
+backcall==0.2.0
+bleach==4.1.0
+certifi==2021.10.8
+cffi==1.15.0
+charset-normalizer==2.0.12
+debugpy==1.5.1
+decorator==5.1.1
+defusedxml==0.7.1
+docutils==0.17.1
+entrypoints==0.4
+executing==0.8.3
+idna==3.3
+imagesize==1.3.0
+ipykernel==6.9.1
+ipython==8.1.1
+ipython-genutils==0.2.0
+ipywidgets==7.6.5
+jedi==0.18.1
+Jinja2==3.0.3
+jsonschema==4.4.0
+jupyter-client==7.1.2
+jupyter-core==4.9.2
+jupyterlab-pygments==0.1.2
+jupyterlab-widgets==1.0.2
+m2r2==0.3.2
+MarkupSafe==2.1.0
+matplotlib-inline==0.1.3
+mistune==0.8.4
+nbclient==0.5.11
+nbconvert==6.4.2
+nbformat==5.1.3
+nbsphinx==0.8.8
+nest-asyncio==1.5.4
+notebook==6.4.8
+numpy==1.22.2
+packaging==21.3
+pandocfilters==1.5.0
+parso==0.8.3
+pexpect==4.8.0
+pickleshare==0.7.5
+prometheus-client==0.13.1
+prompt-toolkit==3.0.28
+ptyprocess==0.7.0
+pure-eval==0.2.2
+pycparser==2.21
+Pygments==2.11.2
+pyparsing==3.0.7
+pyrsistent==0.18.1
+python-dateutil==2.8.2
+pytz==2021.3
+pyzmq==22.3.0
+requests==2.27.1
+scipy==1.8.0
+Send2Trash==1.8.0
+six==1.16.0
+snowballstemmer==2.2.0
+Sphinx==4.4.0
+sphinxcontrib-applehelp==1.0.2
+sphinxcontrib-devhelp==1.0.2
+sphinxcontrib-htmlhelp==2.0.0
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.3
+sphinxcontrib-serializinghtml==1.1.5
+stack-data==0.2.0
+terminado==0.13.2
+testpath==0.6.0
+tornado==6.1
+traitlets==5.1.1
+urllib3==1.26.8
+wcwidth==0.2.5
+webencodings==0.5.1
+webgui-jupyter-widgets==0.2.8
+widgetsnbextension==3.5.2
