+++ .github/workflows/build.yml
@@ -156,6 +156,79 @@ jobs:
           tags: ${{ steps.meta.outputs.tags }}
           #labels: ${{ steps.meta.outputs.labels }}
 
+  docs:
+    needs: [ubuntu,macos]
+    runs-on: ubuntu-latest
+    #if: github.event_name == 'release' && github.event.action == 'created'
+    steps:
+      - uses: actions/checkout@master
+        with:
+          fetch-depth: 0 # otherwise, you will failed to push refs to dest repo
+          submodules: 'True'
+          path: 'NGSTrefftz'
+      - name: install pkgs
+        run: |
+             sudo apt-get update && DEBIAN_FRONTEND="noninteractive" sudo apt-get -y install vim libxmu-dev tk-dev tcl-dev cmake git g++ libglu1-mesa-dev libblas-dev liblapack-dev python3 python3-pip python3-distutils python3-tk libpython3-dev 
+             sudo apt-get install -y software-properties-common
+             sudo add-apt-repository universe
+             sudo add-apt-repository ppa:ngsolve/nightly -y
+             sudo apt-get install ngsolve -y
+             sudo apt-get install npm nodejs pandoc -y
+      - name: pip pkgs
+        run: python3 -m pip list
+      - name: set path
+        run: |
+             echo "NETGENDIR=/usr/bin/" >> $GITHUB_ENV
+      - run: pip install --user -r $GITHUB_WORKSPACE/NGSTrefftz/docs/requirements.txt
+      - name: webgui
+        run: |
+             jupyter nbextension install --user --py widgetsnbextension
+             jupyter nbextension enable --user --py widgetsnbextension
+             jupyter nbextension install --user --py webgui_jupyter_widgets
+             jupyter nbextension enable --user --py webgui_jupyter_widgets            
+      - run: python3 -c "import sys; print(sys.path)"
+      - name: cmake NGSTrefftz
+        run: |
+             cmake -B $GITHUB_WORKSPACE/NGSTrefftz/make -S $GITHUB_WORKSPACE/NGSTrefftz/src/
+             make -C $GITHUB_WORKSPACE/NGSTrefftz/make
+             sudo make -C $GITHUB_WORKSPACE/NGSTrefftz/make install  
+             python3 -c "import sys; print(sys.path)"
+             python3 -c "import ngstrefftz"
+             ln -s /usr/lib/python3/dist-packages/ngsolve /opt/hostedtoolcache/Python/3.10.2/x64/lib/python3.10/site-packages/
+             ln -s /usr/lib/python3/dist-packages/ngstrefftz /opt/hostedtoolcache/Python/3.10.2/x64/lib/python3.10/site-packages/
+             ln -s /usr/lib/python3/dist-packages/ngstents /opt/hostedtoolcache/Python/3.10.2/x64/lib/python3.10/site-packages/
+      - name: jupyter stuff
+        run: |
+             pip install jupyter_contrib_nbextensions 
+             #pip install jupyter_nbextensions_configurator 
+             jupyter contrib nbextension install --user
+             jupyter nbextension install --user --py widgetsnbextension
+             jupyter nbextension enable --user --py widgetsnbextension
+      - name: build tentswebgui
+        run: |
+             cd $GITHUB_WORKSPACE/NGSTrefftz/external_dependencies/ngstents/tentswebgui/ && bash build.sh
+      - name: install tentswebgui
+        run: |
+             cd $GITHUB_WORKSPACE/NGSTrefftz/external_dependencies/ngstents/tentswebgui/ && pip install --user .
+      - name: jupyter tentswebgui
+        run: |
+             jupyter nbextension install --user --py tentswebgui
+             jupyter nbextension enable --user --py tentswebgui
[…]
