### 2022-03-06 8b73c4b
publish gh-page only on release
