### 2022-03-06 PaulSt/NGSTrefftz#8 @PaulSt
Thanks for pointing that out! The documentation for the python API is now available at https://paulst.github.io/NGSTrefftz/
