### 2022-03-06 PaulSt/NGSTrefftz#7 @PaulSt
Thank you for sharing the example, I have now added the guidelines!
