### 2022-03-09 PaulSt/NGSTrefftz#8 @thelfer
@PaulSt, This page has rendering issues: <https://paulst.github.io/NGSTrefftz/notebooks/embTrefftz.html>
