### 2022-03-09 PaulSt/NGSTrefftz#8 @thelfer
@PaulSt I thing that the website <https://paulst.github.io/NGSTrefftz> shall be referenced in the README.md which is displayed when looking at github repository of the project.
