### 2022-03-09 PaulSt/NGSTrefftz#9 @thelfer
Review of the JOSS paper
This issue is created as part of the review of the JOSS paper (openjournals/joss-reviews#4135).

My biggest concern with the paper concerns the following statements:

1. line 19: NGSTrefftz provides a framework to implement Trefftz finite element spaces
2. line 40: The aim of this package is to facilitate research into Trefftz methods and to make them more
accessible to a broad audience.

My feeling is that the `NGSTrefftz` currently  does not provide an appealing introduction to the Trefftz method (not in the paper, but on the website and a reference to it in the paper) that would make the reader want to give it a try and maybe contribute to your project.

I would have expected to find, for example, a introductive text to the method that would be accessible to standard engineer. A minima, a description of the advantages of the Trefftz method (regarding other methods) would help to explain why "research into Trefftz methods" are important.

I have understood that the main drawback of the Trefftz metho is that it requires knowledge of the PDE to solve, but I did not see how I could extend `NGSTrefftz` for my particular PDE (example: linear elasticity or hyperelastictity). Shall I go at the C++ level. Shall I use the "general framework to produce Trefftz spaces implicitly" ?

By that, I do mean that one can not expect the interrested user to go through the provided notebooks (unless you're more than lucky).
