+++ docs/requirements.txt
@@ -30,6 +30,7 @@ jupyterlab-pygments==0.1.2
 jupyterlab-widgets==1.0.2
 m2r2==0.3.2
 MarkupSafe==2.1.0
+matplotlib
 matplotlib-inline==0.1.3
 mistune==0.8.4
 nbclient==0.5.11
