+++ docs/index.rst
@@ -5,6 +5,107 @@
 
 .. mdinclude:: ../README.md
 
+.. .. include:: notebooks/laplace.rst
+
+
+Introduction to Trefftz-DG
+========================================
+
+
+.. jupyter-execute::
+    :hide-code:
+    :hide-output:
+
+    from ngsolve import *
+    from ngstrefftz import *
+    from netgen.geom2d import unit_square
+
+.. jupyter-execute::
+    :hide-output:
+
+    mesh = Mesh(unit_square.GenerateMesh(maxh=.3))
+    fes = trefftzfespace(mesh,order=4,eq="laplace")
+
+.. jupyter-execute::
+    :hide-code:
+    :hide-output:
+
+    exact = exp(x)*sin(y)
+    bndc = exact
+
+    def lap_problem(fes):
+        mesh = fes.mesh
+        order = fes.globalorder
+        alpha = 4
+        n = specialcf.normal(mesh.dim)
+        h = specialcf.mesh_size
+        u = fes.TrialFunction()
+        v = fes.TestFunction()
+
+        jump_u = (u-u.Other())*n
+        jump_v = (v-v.Other())*n
+        mean_dudn = 0.5 * (grad(u)+grad(u.Other()))
+        mean_dvdn = 0.5 * (grad(v)+grad(v.Other()))
+
+        a = BilinearForm(fes,symmetric=True)
+        a += grad(u)*grad(v) * dx \
+            +alpha*order**2/h*jump_u*jump_v * dx(skeleton=True) \
+            +(-mean_dudn*jump_v-mean_dvdn*jump_u) * dx(skeleton=True) \
+            +alpha*order**2/h*u*v * ds(skeleton=True) \
+            +(-n*grad(u)*v-n*grad(v)*u)* ds(skeleton=True)
+
+        f = LinearForm(fes)
+        f += alpha*order**2/h*bndc*v * ds(skeleton=True) \
+             +(-n*grad(v)*bndc)* ds(skeleton=True)
+
+        with TaskManager():
+            a.Assemble()
+            f.Assemble()
+        return (a,f)
+
+.. jupyter-execute::
+
+    errors = []
+    for p in range(1,6):
+        errors.append([])
+        mesh = Mesh(unit_square.GenerateMesh(maxh=1))
+        for h in range(6):
+            fes = trefftzfespace(mesh,order=p,eq="laplace")
+            (a,f) = lap_problem(fes)
+            gfu = GridFunction(fes)
+            with TaskManager():
+                gfu.vec.data = a.mat.Inverse(inverse='sparsecholesky') * f.vec
+            errors[p-1].append(sqrt(Integrate((gfu-exact)**2, mesh)))
+            mesh.Refine()
+    print(errors)
+
+
+.. jupyter-execute::
+    :hide-code:
+
+    import matplotlib.pyplot as plt
+
+    # plt.subplots_adjust(hspace=1)
+
+    # log y axis
+    plt.subplot(121)
+    plt.semilogy()
+    plt.title('semilogy')
+    plt.grid(True)
+
+    # log x and y axis
+    plt.subplot(122)
+    for p in range(1,6):
+        plt.loglog([0.5**i for i in range(6)], errors[p-1],'-o',label=f'p = {p}')
+    ax=plt.gca()
+    ax.invert_xaxis()
+    plt.grid(True)
+    plt.legend()
+    plt.title('loglog base 2 on x')
+
+    plt.show()
+
+
 
 Documentation
 ========================================
