+++ src/CMakeLists.txt
@@ -90,6 +90,9 @@ file(COPY
     ${CMAKE_SOURCE_DIR}/../test/trefftz.py
     ${CMAKE_SOURCE_DIR}/../test/tents.py
      DESTINATION ${CMAKE_BINARY_DIR}/Testing)
-add_test(NAME embtrefftz COMMAND python3 ${CMAKE_BINARY_DIR}/Testing/embt.py -v)
-add_test(NAME trefftz COMMAND python3 ${CMAKE_BINARY_DIR}/Testing/trefftz.py -v)
-add_test(NAME tents COMMAND python3 ${CMAKE_BINARY_DIR}/Testing/tents.py -v)
+add_test(NAME embtrefftz COMMAND python3 -m doctest ${CMAKE_BINARY_DIR}/Testing/embt.py)
+add_test(NAME trefftz COMMAND python3 -m doctest ${CMAKE_BINARY_DIR}/Testing/trefftz.py)
+add_test(NAME tents COMMAND python3 -m doctest ${CMAKE_BINARY_DIR}/Testing/tents.py)
+#WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}
+set_tests_properties(embtrefftz trefftz tents
+    PROPERTIES ENVIRONMENT "PYTHONPATH=${CMAKE_BINARY_DIR}:$ENV{PYTHONPATH}")
