+++ docs/conf.py
@@ -1,3 +1,57 @@
+from os.path import dirname, isdir, join
+import os
+import re
+import subprocess
+def get_version():
+    """
+    Gets the current version number.
+    If in a git repository, it is the current git tag.
+    Otherwise it is the one contained in the PKG-INFO file.
+    """
+    version_re = re.compile('^Version: (.+)$', re.M)
+    d = dirname(__file__)+'/..'
+
+    if isdir(join(d, '.git')):
+        # Get the version using "git describe".
+        cmd = 'git describe --always --tags --match v[0-9]*'.split()
+        try:
+            version = subprocess.check_output(cmd).decode().strip()
+        except subprocess.CalledProcessError:
+            print('Unable to get version number from git tags')
+            exit(1)
+
+        # PEP 386 compatibility
+        if '-' in version:
+            version = '.post'.join(version.split('-')[:2])
+
+        # Don't declare a version "dirty" merely because a time stamp has
+        # changed. If it is dirty, append a ".dev1" suffix to indicate a
+        # development revision after the release.
+        with open(os.devnull, 'w') as fd_devnull:
+            subprocess.call(['git', 'status'],
+                            stdout=fd_devnull, stderr=fd_devnull)
+
+        cmd = 'git diff-index --name-only HEAD'.split()
+        try:
+            dirty = subprocess.check_output(cmd).decode().strip()
+        except subprocess.CalledProcessError:
+            print('Unable to get git index status')
+            exit(1)
+
+        # if dirty != '':
+            # version += '.dev1'
+
+        # strip the v for pypi
+        version = version[1:]
+
+    else:
+        # Extract the version from the PKG-INFO file.
+        with open(join(d, 'PKG-INFO')) as f:
+            version = version_re.search(f.read()).group(1)
+
+    return version
+
+
 # Configuration file for the Sphinx documentation builder.
 #
 # This file only contains a selection of the most common options. For a full
@@ -22,7 +76,7 @@ copyright = '2022, Paul Stocker'
 author = 'Paul Stocker'
 
 # The full version, including alpha/beta/rc tags
-release = 'v0.0.4'
+release = get_version()
 
 
 # -- General configuration ---------------------------------------------------
