### 2022-03-11 76aac15
output ctest on failure
