+++ test/tents.py
@@ -124,7 +124,7 @@ def TestQTrefftz(order, initmesh, t_step,qtrefftz=1):
     ...        initmesh = Mesh(SegMesh(h,0,math.pi))
     ...        TestQTrefftz(order,initmesh,t_step) # doctest:+ELLIPSIS
     0.02...
-    0.0011...
+    0.001...
     ...e-05
     ...e-06
     >>> initmesh = Mesh(unit_square.GenerateMesh(maxh = 0.5))
