### 2022-03-11 7dd2c60
output ctest on failure
