### 2022-03-12 PaulSt/NGSTrefftz#9 @PaulSt
Hi @thelfer,

thank you again for the thorough review. I have updated the website to address these issues.
I put a concise example with convergence plots directly onto the website. Additionally, I put a short text for a starting point into NGSTrefftz.
The notebooks are a common tool in the NGSolve community. There are still a few options to make the more user friendly to be added in the future. Let me know if this resolves the issue for you for now.
