+++ README.md
@@ -47,6 +47,8 @@ You need to have [NGSolve](https://www.ngsolve.com/) installed. To build the pac
 To access the newest features the nightly version of NGSolve works best and lapack >= 3.9 is required.
 
 ## News
+Mar, 2022: NGSTrefftz now has a [website](https://paulst.github.io/NGSTrefftz/)! 🚀
+
 Feb, 2022: If you are using NGSolve nightly releases: [NGSolve@eda758d](https://github.com/NGSolve/ngsolve/commit/eda758d99483888851913d8a5c9aff4d0cbc9cc2) breaks a dependency and [NGSolve@3d52ecd](https://github.com/NGSolve/ngsolve/commit/3d52ecd615f2b7c409219eebaba99288ea19c1bc) produces import issue. Make sure to update ngstrefftz submodules and move to newest ngsolve version, at least [NGSolve@5839a09](https://github.com/NGSolve/ngsolve/commit/5839a09810235a938bd85807d8e638d3a0b6c69d).
 
 Jan, 2022: NGSTrefftz is now available via pip! 🚀
