+++ docs/paper/paper.bib
@@ -131,3 +131,13 @@
  MSC2010 = {74S05 74-02 65N30},
  Zbl = {0962.74071}
 }
+
+@article{freehyte,
+  author = {Ionu\c{t} D. Moldovan and Ildi Cisma\c{s}iu},
+  title = {{FreeHyTE}: a hybrid-{T}refftz finite element platform},
+  journal = {Advances in Engineering Software}
+  volume = {121},
+  pages = {{98-119},
+  year = {2018},
+  doi = {10.1016/j.advengsoft.2018.03.014}
+}
