### 2022-03-15 #4135 @mscroggs
@PaulSt Nice software and nice paper!

I've opened a PR (https://github.com/PaulSt/NGSTrefftz/pull/10) to remove some brackets around references and add a reference to the paper that the FreeHyTe people wrote. Once that PR is resolve, I think this is ready to be accepted
