+++ docs/paper/paper.bib
@@ -135,9 +135,9 @@
 @article{freehyte,
   author = {Ionu\c{t} D. Moldovan and Ildi Cisma\c{s}iu},
   title = {{FreeHyTE}: a hybrid-{T}refftz finite element platform},
-  journal = {Advances in Engineering Software}
+  journal = {Advances in Engineering Software},
   volume = {121},
-  pages = {{98-119},
+  pages = {{98-119}},
   year = {2018},
   doi = {10.1016/j.advengsoft.2018.03.014}
 }
