### 2022-03-18 #4135 @Nikoleta-v3
Hey @PaulSt 👋🏻 Could you please have a look at the following (minor) issues?

### Documentation

On the NGSTrefftz [website](https://paulst.github.io/NGSTrefftz/):

1. The hyperlinks to NGSolve do not work
2. The emojis `:tada::+1:` are not rendered

### Readme

At the installation section you are listing the dependencies and their versions :+1: Could you include hyperlinks to the software’s website, or their installation pages?

This is more of a suggestion: why don’t you make `Papers using the code section` a section instead of a subsection?

### Testing

I can see you have some tests (folder `test`) but I couldn't find any information on how to run the tests. Can you include this information in the documentation? And please point to it in, or just include it under `Contributing to NGSTrefftz`.

I can also see that you are running GitHub actions. Could you also run the test suite as part of the workflow?
