### 2022-03-22 #4135 @Nikoleta-v3
Thank you @PaulSt. Note that you left some commented out [lines](https://github.com/PaulSt/NGSTrefftz/blob/ffe8537814137f59bb0f4d0a31d2527f9b78042e/.github/workflows/build.yml#L75-L81) in the `build.yml` file.

At this point could you:

- [ ] Let me know which tagged release of your software we should use (I am assuming [v0.0.7](https://github.com/PaulSt/NGSTrefftz/releases/tag/v0.0.7)).

I can then move forward with accepting the submission. Let me know if you need any help with any of the above 😄
