+++ docs/conf.py
@@ -78,7 +78,6 @@ author = 'Paul Stocker'
 # The full version, including alpha/beta/rc tags
 release = get_version()
 
-
 # -- General configuration ---------------------------------------------------
 
 # Add any Sphinx extension module names here, as strings. They can be
@@ -105,7 +104,7 @@ run_notebook_export_template = 'basic'  # Default: 'full'
 run_notebook_display_source_links = False  # Default: True
 
 # Whether or not to evaluate the notebooks prior to embedding them
-evaluate_notebooks = False  # Default: True
+evaluate_notebooks = True  # Default: True
 
 # START nbsphinx stuff
 #increase timeout for cell execution, since some files take long to execute
@@ -137,9 +136,9 @@ nbsphinx_prolog = r"""
 .. only:: html
     .. role:: raw-html(raw)
         :format: html
-
 """
 
+nbsphinx_widgets_path = ""
 # END nbsphinx stuff
 
 # The name of the Pygments (syntax highlighting) style to use.
@@ -156,25 +155,6 @@ templates_path = ['_templates']
 # This pattern also affects html_static_path and html_extra_path.
 exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store', 'paper','env']
 
-# LaTeX
-latex_engine = 'xelatex'
-latex_elements = {
-    'fontpkg': r'''
-\setmainfont{DejaVu Serif}
-\setsansfont{DejaVu Sans}
-\setmonofont{DejaVu Sans Mono}
-''',
-    'preamble': r'''
-\usepackage[titles]{tocloft}
-\cftsetpnumwidth {1.25cm}\cftsetrmarg{1.5cm}
-\setlength{\cftchapnumwidth}{0.75cm}
-\setlength{\cftsecindent}{\cftchapnumwidth}
-\setlength{\cftsecnumwidth}{1.25cm}
-''',
-    'fncychap': r'\usepackage[Bjornstrup]{fncychap}',
-    'printindex': r'\footnotesize\raggedright\printindex',
-}
-
 
 # -- Options for HTML output -------------------------------------------------
 
@@ -185,12 +165,23 @@ html_theme = 'alabaster'
 html_theme_options = {
     'github_user': 'PaulSt',
     'github_repo': 'NGSTrefftz',
+    # 'github_banner':True,
+    # 'travis_button':True,
+    'github_button':True,
+    'fixed_sidebar':True,
+    'font_size':10
     # 'analytics_id': 'G-XXXXXXXXXX',  #  Provided by Google in your dashboard
     # 'analytics_anonymize_ip': False,
 }
 html_sidebars = {
-   'index': ['localtoc.html'],
-   '**': [],
+   'index': [
+        'about.html',
+        'localtoc.html',
+        # 'relations.html',
+       ],
+   '**': ['about.html',
+          'localtoc.html',
+       ],
 }
 
 
@@ -199,4 +190,4 @@ html_sidebars = {
 # so a file named "default.css" will overwrite the builtin "default.css".
 html_static_path = ['_static']
 
-html_js_files = ['webgui_jupyter_widgets.js', 'webgui.js', 'tentswebgui.js']
+# html_js_files = ['webgui_jupyter_widgets.js', 'webgui.js', 'tentswebgui.js']
