### 2022-03-29 #4135 @PaulSt
Hi @Nikoleta-v3,
and sorry for the delay!
- [X] I published a release which I would like to use for the software, https://github.com/PaulSt/NGSTrefftz/releases/tag/v0.1.0
- [X] The software is published in Zenodo at [10.5281/zenodo.6394628](https://zenodo.org/record/6394628)
