### 2022-03-30 #4135 @PaulSt
Hi @Nikoleta-v3,
I was just checking if the error was an issue in the paper but locally it compiled. Is it possible that
> @editorialbot set [10.5281/zenodo.6394628](https://zenodo.org/record/6394628) as archive

loops back onto itself? So one would need to call it without the link? :thinking:
