### 2022-03-30 #4135 @PaulSt
>
> Check final proof point_right [openjournals/joss-papers#3099](https://github.com/openjournals/joss-papers/pull/3099)
>
@Nikoleta-v3 is missing as the editor in the proof, everything else looks good to me!
