### 2022-03-30 #4135 @kyleniemeyer
@openjournals/dev I think the editor should be showing up properly in the proof, any idea what's going on there?
