### 2022-03-31 #4135 @xuanxu
> @openjournals/dev I think the editor should be showing up properly in the proof, any idea what's going on there?

Probably this can be fixed by @Nikoleta-v3 filling in *first_name/last_name* fields in her [editor profile page](https://joss.theoj.org/editor_profile).
