### 2022-03-31 #4135 @xuanxu
Ouch, then sounds like there's a bug somewhere in the code preparing the paper. I'm going to try to locate it.
