### 2022-03-31 #4135 @xuanxu
Bug found, it was caused by incorrectly managing usernames with upcased letters.
Fix applied, regenerating the paper should get the right editor info now.
