### 2022-03-31 #4135 @Nikoleta-v3
Thank you very much @xuanxu!

Should I run the `generate pdf command`? will this update the final proofs or should I re-run `@editorialbot recommend-accept`?
