### 2022-03-31 #4135 @PaulSt
Hi @kyleniemeyer,

thank you very much and many thanks to @Nikoleta-v3 for editing!

I was looking at the publication and was surprised to see the quotation mark ` used to style the title in markdown made it into the title on the joss homepage. I don't know how that happened, and my apologies if this was an error on my side. If still possible, it might be nicer to remove it, to make it more uniform with the other papers.
