### 2022-04-01 #4135 @kyleniemeyer
@PaulSt ah, good catch. that is probably not supposed to happen.

@openjournals/dev is this something that can be fixed?
